Add containers and  footers
Rigth now it's harder to control the heigh of the rows for the layout
I have in mind. I put the header outside my grid so I could control
it size better.

The distribution is ok. I have to research how to control the viewport
height to better serve the layout as I intended.